backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.25
email-validator==2.1.0